:bookmark: λcommon v1.7.3: Fixed compilation problem on Linux with the C wrapper.

+++ c_wrapper/tests/test.c
@@ -102,7 +102,7 @@ int main()
 	if (testsPassed != testsCount)
 		return 1;
 
-	printf("\nTests execution time: %lldms", (lc_time_getTimeMillis() - start));
+	printf("\nTests execution time: %lldms", (long long) (lc_time_getTimeMillis() - start));
 
 	return 0;
 }
